pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "fastapi[standard] (>=0.133.1,<0.134.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "langgraph (>=1.0.9,<2.0.0)",
    "langchain (>=1.2.10,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "langchain-huggingface (>=1.2.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "pypdf (>=6.7.3,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "typing-extensions (>=4.15.0,<5.0.0)",
    "langchain-core (>=1.2.16,<2.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "google-generativeai (>=0.8.0,<0.9.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)"
]
